+++ src/app/sites/info-teaching/components/about-futureskills/about-futureskills.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'fs-about-futureskills',
+  templateUrl: './about-futureskills.component.html',
+  styleUrls: ['./about-futureskills.component.scss']
+})
+export class AboutFutureskillsComponent {
+
+}