+++ src/app/sites/info-teaching/components/video-presentation/video-presentation.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'fs-video-presentation',
+  templateUrl: './video-presentation.component.html',
+  styleUrls: ['./video-presentation.component.scss']
+})
+export class VideoPresentationComponent {
+}